Feat: Resolvidos o desafio, implementado os components de sideBar e content.
